+++ src/components/forms/fields/LvefInput.vue
@@ -61,15 +61,6 @@
                     :color="item.color"
                 ></v-radio>
                 </div>
-                <div v-if="fieldData.component == 'v-checkbox-group'">
-                <v-radio
-                    v-for="(item, index) in fieldData.items"
-                    :key="index"
-                    :label="item.label"
-                    :value="item.value"
-                    :color="item.color"
-                ></v-radio>
-                </div>
                 <p class="mb-3 mt-4" v-if="fieldData.component == 'h2'">
                 {{ fieldData.value }}
                 </p>
@@ -206,7 +197,6 @@ export default {
   },
   methods: {
     customFieldUpdate: function(event){
-        console.log('custom event field update')
         this.fieldData.value = event;
     },
     getMasking: function() {
@@ -214,10 +204,8 @@ export default {
             !this.isUndefined(this.fieldData.masking) && 
             this.fieldData.masking !== null
         ) {
-            console.log('MASKING', 'return data triggered', this.fieldData.masking);
             return this.fieldData.masking;
         }
-        console.log('MASKING', 'return false triggered');
         return null;
     },
     fieldValueLength: function(value) {
@@ -273,12 +261,18 @@ export default {
             selectItems.unshift(anyfield);
           }
         }
+        result["item_value"] = field.item_value;
+        result["item_text"] = field.item_text;
         result["items"] = selectItems;
         result["chips"] = field.chips;
         result["data-vv-name"] = "select";
       }
 
-      if (field.type == "lvef-checkbox-group") {
+      if (field.type == "checkbox-group") {
+        if (!this.isUndefined(field.switch)) {
+            result["switch"] = field.switch;
+        }
+
         result["items"] = field.items;
         result["item"] = field.item_value;
         result["data-vv-name"] = "checkbox";
@@ -337,6 +331,94 @@ export default {
 
       if (!this.isUndefined(field.maxlength)) {
         result["maxlength"] = field.maxlength;
+      }
+
+      if (!this.isUndefined(field.appendIcon)) {
+        result["appendIcon"] = field.appendIcon;
+      }
+
+      if (!this.isUndefined(field.backgroundColor)) {
+        result["backgroundColor"] = field.backgroundColor;
+      }
+
+      if (!this.isUndefined(field.error)) {
+        result["error"] = field.error;
+      }
+
+      if (!this.isUndefined(field.errorCount)) {
+        result["errorCount"] = field.errorCount;
+      }
+
+      if (!this.isUndefined(field.errorMessages)) {
+        result["errorMessages"] = field.errorMessages;
+      }
+
+      if (!this.isUndefined(field.falseValue)) {
+        result["falseValue"] = field.falseValue;
+      }
+
+      if (!this.isUndefined(field.flat)) {
+        result["flat"] = field.flat;
+      }
+
+      if (!this.isUndefined(field.hideDetails)) {
+        result["hideDetails"] = field.hideDetails;
+      }
+
+      if (!this.isUndefined(field.hideSpinButtons)) {
+        result["hideSpinButtons"] = field.hideSpinButtons;
+      }
+
+      if (!this.isUndefined(field.hint)) {
+        result["hint"] = field.hint;
+      }
+
+      if (!this.isUndefined(field.inputValue)) {
+        result["inputValue"] = field.inputValue;
+      }
+
+      if (!this.isUndefined(field.inset)) {
+        result["inset"] = field.inset;
+      }
+
+      if (!this.isUndefined(field.dark)) {
+        result["dark"] = field.dark;
+      }
+
+      if (!this.isUndefined(field.light)) {
+        result["light"] = field.light;
+      }
+
+      if (!this.isUndefined(field.messages)) {
+        result["messages"] = field.messages;
+      }
+
+      if (!this.isUndefined(field.persistentHint)) {
+        result["persistentHint"] = field.persistentHint;
+      }
+
+      if (!this.isUndefined(field.prependIcon)) {
+        result["prependIcon"] = field.prependIcon;
+      }
+
+      if (!this.isUndefined(field.success)) {
+        result["success"] = field.success;
+      }
+
+      if (!this.isUndefined(field.successMessages)) {
+        result["successMessages"] = field.successMessages;
+      }
+
+      if (!this.isUndefined(field.trueValue)) {
+        result["trueValue"] = field.trueValue;
+      }
+
+      if (!this.isUndefined(field.validateOnBlur)) {
+        result["validateOnBlur"] = field.validateOnBlur;
+      }
+
+      if (!this.isUndefined(field.display)) {
+        result["display"] = field.display;
       }
 
       if (field.component == "h2") {